Use steepest descent for inversion
Need to change the Jacobian for that. So I needed to stop caching the
Jacobian and Hessian (makes no difference for performance). Performance
if the same as Gauss-Newton. Levemberg-Marquardt is also not very good
because the step search involves more forward modeling (the bottle neck
of this inversion).

Steepest is not very good for regularization. Included example showing
that Gauss-Newton performs better for smoothness.

Get rid of gradient inversion (leave for later paper).