Move profiling code to simple synthetic notebook
Run the profiling and save the raw results as well in a way that can be
loaded by pstats. Extract the execution times of conjugate gradient, dot
products and forward modeling. Generate the Latex table with these
numbers and save it to a tex file for inclusion in the manuscript.

Need to run the entire notebook again.